requirements.txt
amqp==2.5.2
APScheduler==3.6.3
beautifulsoup4==4.9.1
billiard==3.6.3.0
blinker==1.4
bs4==0.0.1
celery==4.4.1
certifi==2020.4.5.2
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-APScheduler==1.11.0
Flask-Mail==0.9.1
Flask-SSLify==0.1.5
idna==2.9
itsdangerous==1.1.0
Jinja2==2.10.1
kombu==4.6.8
MarkupSafe==1.1.1
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2019.1
redis==3.4.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
tzlocal==2.1
urllib3==1.25.9
vine==1.3.0
Werkzeug==1.0.0
